Moved ExpectationUtils in prep for providing access to spring context
